Some final changes that were overlooked when reconciling Thomas introduction of the abSynapse and David's introduction of AEC.

+++ src/gui/MeasMethodDlg.cpp
@@ -490,7 +490,7 @@ bool MeasMethodDlg::generalInit(bool kernelTest)
         }                
     }
 
-    numberOfSamples = params.injLen*params.samplingRate;
+    numberOfSamples = (long) (params.injLen*params.samplingRate);
     samplingPeriod  = 1/params.samplingRate;
 
     // Reinit data storing objects
